requirements.txt
requests==2.31.0
openai==0.27.8
schedule==1.2.0
python-dotenv==1.0.0
